requirements.txt
ruff==0.6.5
pyodbc==5.1.0
sshtunnel==0.4.0
pandas==2.2.2
hatch==1.12.0
google-auth==2.34.0
google-cloud-bigquery==3.25.0
google-cloud-storage==2.18.2
fastparquet==2024.5.0